# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Clinic:
    properties:
      abnnumber:
        type: string
      address:
        type: string
      city:
        type: string
      clinicShare:
        type: integer
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      logoURL:
        type: string
      methodType:
        $ref: '#/definitions/domain.MethodType'
      name:
        type: string
      ownerShare:
        type: integer
      phone:
        type: string
      postcode:
        type: string
      shareType:
        $ref: '#/definitions/domain.ShareType'
      state:
        type: string
      updatedAt:
        type: string
      website:
        type: string
      withHoldingTax:
        type: boolean
    type: object
...
  domain.H:
    additionalProperties: true
    type: object
...
  domain.MethodType:
    enum:
    - NET
    - GROSS
    type: string
    x-enum-varnames:
    - MethodTypeNet
    - MethodTypeGross
...
  domain.ShareType:
    enum:
    - PERCENTAGE
    - FIXED
    type: string
    x-enum-varnames:
    - ShareTypePercentage
    - ShareTypeFixed
...
host: localhost:8080
info:
  contact:
    email: support@acareca.com
    name: API Support
  description: API documentation for ACA RECA Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ACA RECA API
  version: "1.0"
paths:
...
  /clinic:
    get:
      consumes:
      - application/json
      description: Retrieve all clinics the authenticated user is associated with
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Retrieve user's clinics
      tags:
      - Clinic
    post:
      consumes:
      - application/json
      description: Create a new clinic with the given information. The creating user
        is automatically associated as owner.
      parameters:
      - description: Clinic information
        in: body
        name: clinic
        required: true
        schema:
          $ref: '#/definitions/domain.Clinic'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Create a new clinic
      tags:
      - Clinic
  /clinic/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a clinic by ID. User must be associated with the clinic.
      parameters:
      - description: Clinic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Delete a clinic by ID
      tags:
      - Clinic
    get:
      consumes:
      - application/json
      description: Retrieve a clinic by ID. User must be associated with the clinic.
      parameters:
      - description: Clinic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Retrieve a clinic by ID
      tags:
      - Clinic
    put:
      consumes:
      - application/json
      description: Update a clinic by ID. User must be associated with the clinic.
      parameters:
      - description: Clinic ID
        in: path
        name: id
        required: true
        type: string
      - description: Clinic information
        in: body
        name: clinic
        required: true
        schema:
          $ref: '#/definitions/domain.Clinic'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Update a clinic by ID
      tags:
      - Clinic
  /clinic/abn/{abnNumber}:
    get:
      consumes:
      - application/json
      description: Retrieve a clinic by ABN number. User must be associated with the
        clinic.
      parameters:
      - description: ABN number
        in: path
        name: abnNumber
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.H'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/domain.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.H'
      summary: Retrieve a clinic by ABN number
      tags:
      - Clinic
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"